feat: add dynamic titles and analytics page tracking
